requirements.txt
pymongo
python-dotenv
numpy
pandas
certifi
scikit-learn
pyyaml
python-box
-e .